fix: keep private Codex switcher labels distinct
Honor Hide Personal Info in account button titles and tooltips, including email text embedded in workspace labels. Prefix private labels with account numbers and preserve those numbers when labels shorten, without changing selection callbacks or ordinary labels.

Reuse existing redaction and test seams while extracting the Codex switcher view. Includes label, selection and synthetic native regression proof.
